refactor into multiple components
