+++ packages/runtime/src/widgets/Date/Date.tsx
@@ -2,31 +2,22 @@ import {
   type EnsembleAction,
   useRegisterBindings,
 } from "@ensembleui/react-framework";
-import { Form as AntForm, Typography } from "antd";
+import { Form as AntForm } from "antd";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
-import { useState, createContext, useContext, useCallback } from "react";
-import {
-  PickersCalendarHeader,
-  type PickersCalendarHeaderProps,
-} from "@mui/x-date-pickers/PickersCalendarHeader";
-import {
-  PickersActionBar,
-  type PickersActionBarProps,
-} from "@mui/x-date-pickers/PickersActionBar";
-import { Button, IconButton, TextField } from "@mui/material";
+import { useState, createContext, useCallback } from "react";
 import dayjs from "dayjs";
-import CalendarTodayIcon from "@mui/icons-material/CalendarToday";
-import { Edit } from "@mui/icons-material";
 import { toString } from "lodash-es";
-import { WidgetRegistry } from "../registry";
-import type { EnsembleWidgetProps } from "../shared/types";
-import { useEnsembleAction } from "../runtime/hooks/useEnsembleAction";
-import type { FormInputProps } from "./Form/types";
+import { WidgetRegistry } from "../../registry";
+import type { EnsembleWidgetProps } from "../../shared/types";
+import { useEnsembleAction } from "../../runtime/hooks/useEnsembleAction";
+import type { FormInputProps } from "../Form/types";
+import { CalendarHeader } from "./CalendarHeader";
+import { ActionBar } from "./ActionBar";
 
-const DateHeaderFormat = "ddd, MMM DD";
-const DateDisplayFormat = "MM/DD/YYYY";
+export const DateHeaderFormat = "ddd, MMM DD";
+export const DateDisplayFormat = "MM/DD/YYYY";
 
 type DateProps = {
   initialValue?: string;
@@ -62,12 +53,12 @@ export const Date: React.FC<DateProps> = (props) => {
   const [isCalendarOpen, setIsCalendarOpen] = useState(false);
   const [enteredDate, setEnteredDate] = useState("");
   const [errorText, setErrorText] = useState("");
+  const action = useEnsembleAction(props.onChange);
 
   const { values } = useRegisterBindings({ ...props, value }, props.id, {
     setValue,
   });
 
-  const action = useEnsembleAction(props.onChange);
   const onChangeCallback = useCallback(
     (date: string) => {
       if (!action) {
@@ -124,6 +115,7 @@ export const Date: React.FC<DateProps> = (props) => {
           <DatePicker
             label={value ? "" : props?.hintText || "Select a date"}
             value={value ? dayjs(value) : ""}
+            disabled={props?.enabled === false}
             closeOnSelect={false}
             open={openPicker}
             maxDate={props.lastDate ? dayjs(props.lastDate) : undefined}
@@ -132,8 +124,8 @@ export const Date: React.FC<DateProps> = (props) => {
             onAccept={onDateAccept}
             onChange={onDateChange}
             slots={{
-              calendarHeader: CustomCalendarHeader,
-              actionBar: CustomActionBar,
+              calendarHeader: CalendarHeader,
+              actionBar: ActionBar,
             }}
             slotProps={{
               actionBar: {
@@ -147,7 +139,10 @@ export const Date: React.FC<DateProps> = (props) => {
                 InputLabelProps: {
                   shrink: false,
                 },
-                onClick: () => setOpenPicker(true),
+                onClick:
+                  props?.enabled === false
+                    ? undefined
+                    : () => setOpenPicker(true),
                 error: false,
                 color: "success",
               },
@@ -165,6 +160,7 @@ export const Date: React.FC<DateProps> = (props) => {
                     overflow: "visible !important",
                     height: "unset !important",
                     maxHeight: "unset !important",
+                    backgroundColor: "rgb(239, 223, 240)",
                   },
                   "& .MuiPickersFadeTransitionGroup-root.MuiDateCalendar-viewTransitionContainer":
                     {
@@ -187,6 +183,7 @@ export const Date: React.FC<DateProps> = (props) => {
                   visibility: `${!value ? "hidden" : "visible"} !important`,
                 },
               "& .MuiDateCalendar-root": {
+                backgroundColor: "rgb(239, 223, 240)",
                 overflow: "visible !important",
                 height: "unset !important",
                 maxHeight: "unset !important",
@@ -204,159 +201,3 @@ export const Date: React.FC<DateProps> = (props) => {
 };
 
 WidgetRegistry.register("Date", Date);
-
-const CustomCalendarHeader: React.FC<PickersCalendarHeaderProps<any>> = (
-  props,
-) => {
-  const context = useContext(DatePickerContext);
-  const [isCalendarOpen, setIsCalendarOpen] = useState(context?.isCalendarOpen);
-
-  const toggleCalendar = (): void => {
-    context?.setIsCalendarOpen!(!context?.isCalendarOpen);
-    setIsCalendarOpen(!isCalendarOpen);
-  };
-
-  const onDateChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ): void => context?.setEnteredDate?.(e.target.value);
-
-  const getHeaderDate = (): string => {
-    const enteredDate = dayjs(context?.enteredDate)?.format(DateHeaderFormat);
-
-    return enteredDate === "Invalid Date"
-      ? context?.value?.format(DateHeaderFormat) || ""
-      : enteredDate;
-  };
-
-  return (
-    <>
-      <div
-        style={{
-          backgroundColor: "rgb(82, 178, 145)",
-        }}
-      >
-        <h6
-          style={{
-            margin: "18px 0 18px 24px",
-            color: "white",
-          }}
-        >
-          SELECT DATE
-        </h6>
-        <Button
-          style={{
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "space-between",
-            backgroundColor: "rgb(82, 178, 145)",
-            borderColor: "rgb(82, 178, 145)",
-            width: "calc(100% - 32px)",
-            marginLeft: "16px",
-          }}
-        >
-          <Typography.Text
-            style={{
-              color: "white",
-              fontSize: "20px",
-              fontWeight: "bold",
-            }}
-          >
-            {getHeaderDate()}
-          </Typography.Text>
-          <IconButton onClick={toggleCalendar}>
-            {isCalendarOpen ? (
-              <Edit
-                sx={{
-                  color: "white",
-                }}
-              />
-            ) : (
-              <CalendarTodayIcon
-                sx={{
-                  color: "white",
-                }}
-              />
-            )}
-          </IconButton>
-        </Button>
-      </div>
-
-      <div>
-        {context?.isCalendarOpen ? (
-          <PickersCalendarHeader
-            {...props}
-            sx={{
-              ...props.sx,
-              "& .MuiPickersCalendarHeader-root": {
-                marginTop: "0px",
-              },
-            }}
-          />
-        ) : (
-          <TextField
-            value={context?.enteredDate}
-            placeholder={DateDisplayFormat.toLowerCase()}
-            InputLabelProps={{ shrink: true }}
-            error={Boolean(context?.errorText) && context?.errorText !== ""}
-            label="Enter Date"
-            helperText={context?.errorText}
-            variant="standard"
-            onChange={onDateChange}
-            sx={{
-              margin: "24px",
-            }}
-          />
-        )}
-      </div>
-    </>
-  );
-};
-
-const CustomActionBar: React.FC<PickersActionBarProps> = (props) => {
-  const context = useContext(DatePickerContext);
-
-  const onSelectAction = (): void => {
-    const dateToCheck = dayjs(context?.enteredDate);
-
-    const isInvalidFormat = context?.isCalendarOpen
-      ? false
-      : dateToCheck.toString() === "Invalid Date" ||
-        !isDateValid(context?.enteredDate);
-    const isOutOfDateRange =
-      (context?.firstDate &&
-        dateToCheck?.isBefore(dayjs(context?.firstDate))) ||
-      (context?.lastDate && dateToCheck?.isAfter(dayjs(context?.lastDate)));
-
-    if (isOutOfDateRange || isInvalidFormat) {
-      let errorMessage = "";
-
-      if (isOutOfDateRange) errorMessage = "Out of range.";
-      else if (isInvalidFormat) errorMessage = "Invalid format.";
-
-      context?.setErrorText?.(errorMessage);
-    } else {
-      context?.setValue?.(dayjs(context?.enteredDate));
-      context?.setErrorText?.("");
-      props.onAccept();
-    }
-  };
-
-  return (
-    <PickersActionBar
-      {...props}
-      onAccept={onSelectAction}
-      sx={{
-        ...props.sx,
-        "& .MuiPickersLayout-actionBar, .MuiButton-root": {
-          color: "rgb(82, 178, 145)",
-          fontSize: "12px",
-        },
-      }}
-    />
-  );
-};
-
-// returns true if date is in mm/dd/yyyy format
-const isDateValid = (date?: string): boolean =>
-  Boolean(date) &&
-  /^(?:0[1-9]|1[0-2])\/(?:0[1-9]|[12][0-9]|3[01])\/\d{4}$/.test(date!);
